pyproject.toml
[project]
name = "fast-whisper"
version = "0.1.0"
description = "Audio transcription service using FastAPI and faster-whisper"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "faster-whisper>=1.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []

